requirements.txt
phi
openai
groq
duckduckgo-search
pytest
pylint
black
streamlit
fastapi
uvicorn
docker
pydantic
google-generativeai

